chore: renaming/restructuring for ai letter gen and templates
