--- internal/api/utils.go
@@ -1,5 +1,0 @@
-package api
-
-func startsWith(s, prefix string) bool {
-	return len(s) >= len(prefix) && s[:len(prefix)] == prefix
-}
