+++ internal/letters/templates/consumer-protection-professional-medium.md
@@ -1,0 +1,36 @@
+---
+Subject: Support for Consumer Protection Legislation - Constituent Request
+Themes: ["consumer protection", "corporate accountability", "fair marketplace"]
+Tone: professional
+Length: medium
+MinWords: 250
+MaxWords: 450
+---
+
+Dear {{.Representative.Title}} {{.Representative.Name}},
+
+I am writing as your constituent from {{.Constituent.ZipCode}} to express my strong support for robust consumer protection measures that ensure fair treatment of {{.Representative.State}} residents in the marketplace.
+
+{{.Advocacy.SpecificConcern}} This issue directly impacts families and individuals across our state who deserve protection from predatory business practices and unfair treatment by corporations that prioritize profits over consumer welfare.
+
+Consumer protection is fundamental to a healthy economy and a fair society. When businesses engage in deceptive practices, hidden fees, unfair contract terms, or fraudulent behavior, it undermines trust in the marketplace and can cause significant financial harm to working families.
+
+I respectfully urge you to {{.Advocacy.RequestedAction}}. Strong consumer protections benefit not only individual consumers but also legitimate businesses that compete fairly and honestly. When we eliminate bad actors from the marketplace, we create a level playing field that rewards quality and integrity.
+
+Effective consumer protection measures should include:
+- Clear disclosure requirements so consumers can make informed decisions
+- Strong enforcement mechanisms with meaningful penalties for violations
+- Accessible avenues for consumers to seek redress when they are harmed
+- Protection for vulnerable populations who may be targeted by predatory practices
+
+Your support for comprehensive consumer protection legislation would demonstrate your commitment to standing with working families against powerful corporate interests that seek to take advantage of consumers.
+
+I believe this issue transcends partisan politics and reflects our shared values of fairness, honesty, and protecting those who may lack the resources to protect themselves against corporate misconduct.
+
+Thank you for considering this important matter. I hope I can count on your leadership to advance meaningful consumer protection measures that will benefit all {{.Representative.State}} residents.
+
+I look forward to your response and to seeing your continued advocacy for consumer rights.
+
+Respectfully,
+{{.Constituent.Name}}
+{{.Constituent.ZipCode}}
