+++ internal/letters/templates/privacy-passionate-long.md
@@ -1,0 +1,36 @@
+---
+Subject: Urgent Action Needed: Protect {{.Representative.State}} Residents from Data Privacy Violations
+Themes: ["data privacy protection", "consumer rights", "corporate accountability"]
+Tone: passionate
+Length: long
+MinWords: 400
+MaxWords: 600
+---
+
+Dear {{.Representative.Title}} {{.Representative.Name}},
+
+My name is {{.Constituent.Name}}, and I am writing to you as a deeply concerned constituent from {{.Constituent.ZipCode}}. The erosion of data privacy rights in our country has reached a critical point, and I urgently need your leadership to protect the fundamental rights of {{.Representative.State}} residents.
+
+{{.Advocacy.SpecificConcern}} This is not just a technical issue—it represents a fundamental violation of our right to privacy and personal autonomy. Every day, corporations collect vast amounts of personal data from Americans without meaningful consent, often selling this information to the highest bidder with little regard for the consequences to individuals and families.
+
+The current situation is unacceptable. Our personal information—including our locations, communications, purchasing habits, health data, and even our children's information—is being harvested, analyzed, and monetized by companies that prioritize profit over privacy. This data collection often happens without our knowledge and certainly without our meaningful consent.
+
+We have seen the devastating consequences of unchecked data collection: identity theft, financial fraud, stalking, discrimination, and the erosion of democratic discourse through targeted misinformation campaigns. The time for half-measures and voluntary industry self-regulation has passed.
+
+I am calling on you to {{.Advocacy.RequestedAction}}. This is not a partisan issue—it is about protecting the fundamental rights of your constituents regardless of their political affiliation. Americans across the political spectrum support stronger privacy protections.
+
+I urge you to consider the following specific actions:
+- Support comprehensive federal privacy legislation that gives individuals real control over their data
+- Advocate for meaningful penalties for companies that violate privacy rights
+- Ensure that privacy protections include strong enforcement mechanisms
+- Protect vulnerable populations, especially children, from predatory data collection practices
+
+Your constituents are counting on you to stand up to powerful corporate interests and fight for our digital rights. The technology industry has shown repeatedly that it cannot be trusted to police itself when it comes to privacy protection.
+
+I respectfully but passionately urge you to make data privacy protection a priority in your legislative agenda. The rights and safety of {{.Representative.State}} residents depend on strong leadership from representatives like you.
+
+Thank you for your time and consideration. I look forward to your swift action on this critical issue.
+
+Sincerely,
+{{.Constituent.Name}}
+Concerned Constituent from {{.Constituent.ZipCode}}
