+++ internal/letters/templates/privacy-professional-short.md
@@ -1,0 +1,24 @@
+---
+Subject: {{.Representative.Title}} {{.Representative.Name}}, Please Support Data Privacy Protection
+Themes: ["data privacy protection", "consumer rights"]
+Tone: professional
+Length: short
+MinWords: 150
+MaxWords: 300
+---
+
+Dear {{.Representative.Title}} {{.Representative.Name}},
+
+My name is {{.Constituent.Name}}, and I am a constituent from {{.Constituent.ZipCode}} in your district. I am writing to urge your support for stronger data privacy protections that safeguard the rights of {{.Representative.State}} residents.
+
+{{.Advocacy.SpecificConcern}} This issue affects millions of Americans who deserve control over their personal information and transparency about how companies collect, use, and share their data.
+
+I respectfully request that you {{.Advocacy.RequestedAction}}. Data privacy is not a partisan issue—it's about protecting consumers from corporate overreach and ensuring individuals have meaningful choice in how their personal information is handled.
+
+Your leadership on this critical issue would make a significant difference for your constituents and set an important precedent for protecting digital rights.
+
+Thank you for your time and consideration.
+
+Respectfully,
+{{.Constituent.Name}}
+Constituent from {{.Constituent.ZipCode}}
